requirements.txt
aiogram==3.15.0
yt-dlp>=2025.1.26
python-dotenv==1.0.1
aiohttp==3.9.5
